HTML and CSS Base Structure
Here is my finished prototype for the HTML and CSS structure of the
site.  To make a page, just include <?php include "header.php"; ?> at
the top and <? php include "footer.php"; ?> at the bottom.  In between
those tags, just act as if you are between the <body></body> tags in an
html document.  The session_start(); is auto-included on every page
through including the header.php file.