Activation page qol/bug fix
Made is so the activation page will no longer throw an error when you
login to an un-activated account (because it doesn't have the code from
the email set in the get).  Instead, it will now be blank if the user
logs in.  However, if the user clicks the link, the session will
remember the code for the user as long as they are logged in (in case
they try to access other pages before finishing activation).  The user
also no longer has access to the survey pages before activation.